feat: overflow protection, hardcoding removal, dead code cleanup
- Add SafeMath (game) and OverflowGuard (simulator) for safe math ops
- Promote Gold/Damage/HP from int to long across 19 files
- Remove hardcoded defaults from GameConfig, BossDropConfig, GoldenGoblinConfig
- Move ComboTracker constants to config/GameData.json
- Delete unused StatFormulas.Generated.cs and Core/Formulas/StatFormulas.cs
- Remove duplicate Simulator/config/ (uses root config/ instead)
- Update Simulator and Core for new overflow-safe patterns

+++ Controls/DamagePopup.cs
@@ -11,7 +11,7 @@ namespace DeskWarrior.Controls
     /// </summary>
     public class DamagePopup : TextBlock
     {
-        public DamagePopup(int damage, bool isCritical = false)
+        public DamagePopup(long damage, bool isCritical = false)
         {
             // 데미지 0일 때 특별 표시
             if (damage == 0)
